requirements.txt
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.0
scipy>=1.7.0
Pillow>=8.3.0
wandb>=0.12.0
einops>=0.6.0
datasets>=2.0.0
tqdm>=4.60.0
